+++ test/Libraries/RevitIntegrationTests/ViewTests.cs
@@ -177,5 +177,167 @@ namespace RevitSystemTests
             var scale = GetPreviewValue("15c8e8f6bffb46bfad5a65900114dc5f");
             Assert.AreEqual(expectedScale, scale);
         }
+
+        [Test]
+        [TestModel(@".\SampleModel.rvt")]
+        public void CanGetSetDiscipline()
+        {
+            // Arrange
+            string samplePath = Path.Combine(workingDirectory, @".\View\CanGetSetDiscipline.dyn");
+            string testPath = Path.GetFullPath(samplePath);
+            
+            // Act
+            ViewModel.OpenCommand.Execute(testPath);
+            RunCurrentModel();
+
+            // Assert
+            var discipline = GetPreviewValue("60c8ce7365214955967367cf21b85762");
+            Assert.AreEqual("Structural", discipline);
+        }
+
+        [Test]
+        [TestModel(@".\SampleModel.rvt")]
+        public void CanGetSetDisplayStyle()
+        {
+            // Arrange
+            string samplePath = Path.Combine(workingDirectory, @".\View\CanGetSetDisplayStyle.dyn");
+            string testPath = Path.GetFullPath(samplePath);
+
+            // Act
+            ViewModel.OpenCommand.Execute(testPath);
+            RunCurrentModel();
+
+            // Assert
+            var displayStyle = GetPreviewValue("269cd16b16ee4b80a8fe6195374c884a");
+            Assert.AreEqual("Shading", displayStyle);
+        }
+
+        [Test]
+        [TestModel(@".\SampleModel.rvt")]
+        public void CanGetSetSketchPlane()
+        {
+            // Arrange
+            string samplePath = Path.Combine(workingDirectory, @".\View\CanGetSetSketchPlane.dyn");
+            string testPath = Path.GetFullPath(samplePath);
+
+            // Act
+            ViewModel.OpenCommand.Execute(testPath);
+            RunCurrentModel();
+
+            // Assert
+            var sketchPlane = GetPreviewValue("bc12df8c2efc44d19a208a83b16e1709") as SketchPlane;
+            var plane = sketchPlane.Plane;
+            Assert.AreEqual(0.000, plane.Origin.X, Tolerance);
+            Assert.AreEqual(0.000, plane.Origin.Y, Tolerance);
+            Assert.AreEqual(2.000, plane.Origin.Z, Tolerance);
+        }
+
+        [Test]
+        [TestModel(@".\SampleModel.rvt")]
+        public void CanGetSetPartsVisibility()
+        {
+            // Arrange
+            string samplePath = Path.Combine(workingDirectory, @".\View\CanGetSetPartsVisibility.dyn");
+            string testPath = Path.GetFullPath(samplePath);
+
+            // Act
+            ViewModel.OpenCommand.Execute(testPath);
+            RunCurrentModel();
+
+            // Assert
+            var partsVisibility = GetPreviewValue("22843a82efe942b2b89437e4a115e69a");
+            Assert.AreEqual("ShowPartsAndOriginal", partsVisibility);
+        }
+
+        [Test]
+        [TestModel(@".\element.rvt")]
+        public void CanGetSetCategoryOverridesAndHidden()
+        {
+            // Arrange
+            string samplePath = Path.Combine(workingDirectory, @".\View\CanGetSetCategoryOverridesAndHidden.dyn");
+            string testPath = Path.GetFullPath(samplePath);
+
+            // Act
+            ViewModel.OpenCommand.Execute(testPath);
+            RunCurrentModel();
+
+            // Assert
+            var categoryOverrides = GetPreviewValue("66dc72c413124c57a8416d0f97bca1de") as Revit.Filter.OverrideGraphicSettings;
+            Assert.IsNotNull(categoryOverrides);
+
+            Assert.IsTrue((bool)GetPreviewValue("c07c8e0e57ab42769fa86b8df7b6056c"));
+        }
+
+        [Test]
+        [TestModel(@".\element.rvt")]
+        public void HideElementTemporary()
+        {
+            // Arrange
+            string samplePath = Path.Combine(workingDirectory, @".\View\HideElementTemporary.dyn");
+            string testPath = Path.GetFullPath(samplePath);
+
+            // Act
+            ViewModel.OpenCommand.Execute(testPath);
+            RunCurrentModel();
+
+            // Assert
+            var view = GetPreviewValue("1fe4f90c6cc044bd91bee7a61258f83b") as Revit.Elements.Views.View;
+            Assert.IsNotNull(view);
+            
+            Assert.IsFalse((bool)GetPreviewValue("0c1c543858f343b89d659f2b537b66dc"));
+        }
+
+        [Test]
+        [TestModel(@".\element.rvt")]
+        public void HideCategoryTemporary()
+        {
+            // Arrange
+            string samplePath = Path.Combine(workingDirectory, @".\View\HideCategoryTemporary.dyn");
+            string testPath = Path.GetFullPath(samplePath);
+
+            // Act
+            ViewModel.OpenCommand.Execute(testPath);
+            RunCurrentModel();
+
+            // Assert
+            var view = GetPreviewValue("b1da33f02c9649bc958299b723222a2c") as Revit.Elements.Views.View;
+            Assert.IsNotNull(view);
+
+            Assert.IsFalse((bool)GetPreviewValue("3bd57778cfb14469bbacd88a81b59341"));
+        }
+
+        [Test]
+        [TestModel(@".\element.rvt")]
+        public void IsolateCategoriesTemporary()
+        {
+            // Arrange
+            string samplePath = Path.Combine(workingDirectory, @".\View\IsolateCategoriesTemporary.dyn");
+            string testPath = Path.GetFullPath(samplePath);
+
+            // Act
+            ViewModel.OpenCommand.Execute(testPath);
+            RunCurrentModel();
+
+            // Assert
+            var view = GetPreviewValue("4cf1acfe08fa406b9fa6977185ef07ac") as Revit.Elements.Views.View;
+            Assert.IsNotNull(view);
+        }
+
+        [Test]
+        [TestModel(@".\element.rvt")]
+        public void IsolateElementsTemporary()
+        {
+            // Arrange
+            string samplePath = Path.Combine(workingDirectory, @".\View\IsolateElementsTemporary.dyn");
+            string testPath = Path.GetFullPath(samplePath);
+
+            // Act
+            ViewModel.OpenCommand.Execute(testPath);
+            RunCurrentModel();
+
+            // Assert
+            var view = GetPreviewValue("f516cb1302f549e3b5eb62b80a7688a7") as Revit.Elements.Views.View;
+            Assert.IsNotNull(view);
+        }
     }
 }